Make the duplicate removal keep shallower files
What I think usually happens is that duplicates are either stored in the
same folder with different names (maybe the OS adds (1) at the end). Or
users try to put all pictures in the same place and when prompted about
pictures with the same filename they decide they don't know if they are
real duplicates so they decide to put them inside a directory.

For this second usecase, we now try to keep the original files that are
more shallow (just if they appear to have the same name). Then if a
folder was full of duplicates, it will be left empty after removal.
These empty folders can then be easily cleaned up.