Complete implementation of file level de duplication
Before buckets were assumed to be all equal or all different, buckets
were not thoroughly checked for duplicate subsets in them. This is now
correctly implemented by sorting bucket's content, comparing the file's
content.

Also added a CLI option so users can pass a removal substring. If this
string is found in the path of a duplicate file, this file will be
preffered for removal over a file with a path that does not contaín the
string. Useful when there is a folder whose content has a very high
probability of being duplicated somewhere else, in such case users would
like to exclusively remove from it and leave the other files somewhere
else.